fix: Get balance query lint errors
